Subatomic openshift functions encapsulated in a SubatomicOpenshiftService class namespace.
